extracted construct_podtemplatespec.go
